Nuke this.
It was an interesting experiment but I think it's probably going to end
up being a sticky situation when it comes to licensing, and it
imbalances the Twilight Forest in ways that I didn't initially expect it
would.

As I don't currently require those resources (although I had planned to)
to be renewable and usable in my modpack, I've decided to just remove
them completely.

It was a learning experience though!